pyproject.toml
[project]
name = "germedbench"
version = "0.1.0"
description = "Open benchmark framework for evaluating LLMs on German clinical texts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0.0",
    "openai>=2.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/germedbench"]
